commit for compile

+++ model/STM32IotNodeTemperature.cpp
@@ -93,7 +93,7 @@ int STM32IotNodeTemperature::requestUpdate()
  float Data;
  if ( ( ( TEMPERATURE_Drv_t* ) DrvContext.pVTable )->Get_Temp( &DrvContext, &Data ) == COMPONENT_OK )
  {
-  sample = Data * 10;
+  sample = ( uint16_t ) ( Data * 10.0 );
   return DEVICE_OK;
  }
  return DEVICE_I2C_ERROR;
